symlink only if not already there

+++ Makefile
@@ -3,4 +3,4 @@ install: *
 	git config --global core.excludesfile ~/.gitignore
 	git config --global color.ui true
 	mkdir -p ~/bin
-	ln -s /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl ~/bin/subl
+	if [ ! -f ~/bin/subl ]; then ln -s /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl ~/bin/subl; fi;
